Organised comparators and improved their javadoc.
