+++ client/src/main/java/org/evosuite/ga/comparators/PreferenceSortingComparator.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with EvoSuite. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.evosuite.ga.metaheuristics.mosa.comparators;
+package org.evosuite.ga.comparators;
 
 import java.util.Comparator;
 
@@ -23,8 +23,9 @@ import org.evosuite.ga.Chromosome;
 import org.evosuite.ga.FitnessFunction;
 
 /**
- * This class implements a <code>Comparator</code> (a method for comparing <code>Chromosomes</code> objects) 
- * based on the dominance test and considering the specified test goals only.
+ * This class implements a <code>Comparator</code> (a method for comparing <code>Chromosomes</code>
+ * objects) based on the fitness value of two chromosome objects and considering the specified test
+ * goals only.
  * 
  * @author Annibale Panichella
  */
@@ -35,27 +36,28 @@ public class PreferenceSortingComparator<T extends Chromosome> implements Compar
 	/**
 	 * Constructor
 	 *
-	 * @param goal
+	 * @param goal a {@link org.evosuite.ga.FitnessFunction} object
 	 */
 	public PreferenceSortingComparator(FitnessFunction<T> goal) {
 		this.objective = goal;
 	}
 
 	/**
-	 * Compares two test cases focusing only on the goals in {@link MOSADominanceComparator#objectives}".
+	 * Compare the fitness value of two chromosome objects focusing only on one goal.
 	 * 
-	 * @param object1
-	 *            Object representing the first test cases.
-	 * @param object2
-	 *            Object representing the second test cases.
-	 * @return -1, or 0, or 1 if object1 dominates object2, both are non-dominated, or solution1 is dominated by solution2, respectively.
+	 * @param object1 a {@link org.evosuite.ga.Chromosome} object
+	 * @param object2 a {@link org.evosuite.ga.Chromosome} object
+	 * @return -1 if fitness value of object1 is lower than the fitness value of object2, 0 if the fitness
+	 *         value of both objects is equal, or 1 if fitness value of object1 is higher than the fitness
+	 *         value of object2.
 	 */
 	@SuppressWarnings("unchecked")
 	public int compare(Object object1, Object object2) {
-		if (object1 == null)
+		if (object1 == null) {
 			return 1;
-		else if (object2 == null)
+		} else if (object2 == null) {
 			return -1;
+		}
 
 		T solution1 = (T) object1;
 		T solution2 = (T) object2;
@@ -63,15 +65,13 @@ public class PreferenceSortingComparator<T extends Chromosome> implements Compar
 		double value1, value2;
 		value1 = solution1.getFitness(this.objective);
 		value2 = solution2.getFitness(this.objective);
-		if (value1 < value2)
+		if (value1 < value2) {
 			return -1;
-		else if (value1 > value2)
+		} else if (value1 > value2) {
 			return +1;
-		else {
+		} else {
 			return solution1.compareSecondaryObjective(solution2);
 		}
-
-	} // compare
-
-} // PreferenceCriterion
+	}
+}
